requirements.txt
crc16==0.1.1
intelhex==2.1
pyserial==3.2.1
pyyaml==3.12
